captain_console/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
django-countries==6.1.2
django-credit-cards==0.4.1
idna==2.9
importlib-metadata==1.6.0
jeepney==0.4.3
keyring==21.2.1
keyrings.alt==1.3
psycopg2==2.8.5
pyasn1==0.4.8
pycparser==2.20
pytz==2019.3
pywin32-ctypes==0.2.0
pyxdg==0.25
requests==2.23.0
SecretStorage==3.1.2
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==3.1.0
